fix: removed warning (size_t -> long)

+++ src/searcher.cpp
@@ -79,7 +79,7 @@ std::vector<Searcher::Print> Searcher::collectPrints( const sys_string& path, co
 #endif
 
     // parse file for newlines until last hit
-    size_t stop = hits.back().second - content.cbegin();
+    long stop = hits.back().second - content.cbegin();
     utils::Lines lines = utils::parseContent( content.data(), content.size(), stop );
 
     std::vector<Hit>::const_iterator hit = hits.cbegin();
